Tests code refactor (#3033)
